fix move. now pl->x and pl->y are double type

+++ cast_horisontal_vertical.c
@@ -7,14 +7,14 @@ double	cast_horisontal(t_main *M, t_hit *hit, double ang)
 
 	sign = 1;
 	hit->side = &M->mp->wall_W;
-	newX = (floor((double)hit->pl_x / 32) + 1) * 32;
+	newX = (floor(hit->pl_x / 32) + 1) * 32;
 	if (ang == 180.0)
 	{
 		sign = -1;
 		hit->side = &M->mp->wall_E;
 		newX = (floor((double)hit->pl_x / 32) * 32) - 0.000001;
 	}
-	while (newX >= 0.0 && (int)newX / 32 < M->map_W[hit->pl_y / 32])
+	while (newX >= 0.0 && (int)newX / 32 < M->map_W[(int)(hit->pl_y / 32)])
 	{
 		if (M->map[(int)(hit->pl_y / 32)][(int)(newX / 32)] == '1')
 			break;
@@ -31,12 +31,12 @@ double	cast_vertical(t_main *M, t_hit *hit, double ang)
 
 	sign = 1;
 	hit->side = &M->mp->wall_N;
-	newY = 32.0 * floor((double)hit->pl_y / 32.0) + 32.0;
+	newY = 32.0 * floor(hit->pl_y / 32) + 32;
 	if (ang == 270.0)
 	{
 		sign = -1;
 		hit->side = &M->mp->wall_S;
-		newY = 32.0 * floor((double)hit->pl_y / 32.0) - 0.000001;
+		newY = 32.0 * floor(hit->pl_y / 32.0) - 0.000001;
 	}
 	while (newY >= 0 && (int)newY / 32 < M->map_H)
 	{
